test: 테스트 설정 수정
- tsd 테스트 설정
- jest 테스트 설정 수정
