+++ __tests__/type/util.test-d.ts
@@ -1,0 +1,6 @@
+import { expectType } from 'tsd';
+import _ from '../../dist/';
+
+expectType<boolean>(_.isNil(null));
+expectType<boolean>(_.isNumber(123));
+expectType<boolean>(_.isFunction(() => {}));
