+++ jest.config.js
@@ -1,4 +1,5 @@
 const config = {
+  testPathIgnorePatterns: ['/node_modules/', '__tests__/type/'],
   transform: {
     '\\.ts$': [
       'ts-jest',
